LayoutTests/fast/events/before-unload-return-string-conversion-expected.txt

CONFIRM NAVIGATION: PASS
Tests that the value returned from a beforeunload event handler gets converted to a string, even if the returnValue property was already set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.defaultPrevented is true
PASS event.returnValue is "PASS"
PASS toStringCalled is true
PASS successfullyParsed is true

TEST COMPLETE
